+++ service-system/src/main/java/jp/co/atguigu/authsystem/service/RoleService.java
@@ -44,5 +44,5 @@ public interface RoleService {
 	 * @param queryVo  クエリVO
 	 * @return Pagination<Role>
 	 */
-	Pagination<Role> pagination(Long pageNum, Long pageSize, RoleQueryVo queryVo);
+	Pagination<Role> pagination(Integer pageNum, Integer pageSize, RoleQueryVo queryVo);
 }
